chore: add utils class
