+++ packages/relay/src/utils.ts
@@ -1,0 +1,38 @@
+/*-
+ *
+ * Hedera JSON RPC Relay
+ *
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import constants from './lib/constants';
+
+export class Utils {
+  public static addPercentageBufferToGasPrice = (gasPrice: number): number => {
+    // converting to tinybar and afterward to weibar again is needed
+    // in order to handle the possibility of an invalid floating number being calculated as a gas price
+    // e.g.
+    //   current gas price = 126
+    //   buffer = 10%
+    //   buffered gas price = 126 + 12.6 = 138.6 <--- invalid tinybars
+    gasPrice +=
+      Math.round(
+        (gasPrice / constants.TINYBAR_TO_WEIBAR_COEF) * (Number(process.env.GAS_PRICE_PERCENTAGE_BUFFER || 0) / 100),
+      ) * constants.TINYBAR_TO_WEIBAR_COEF;
+
+    return gasPrice;
+  };
+}
