+++ packages/relay/tests/lib/formatters.spec.ts
@@ -20,7 +20,6 @@
 
 import { expect } from 'chai';
 import {
-  addPercentageBufferToGasPrice,
   hexToASCII,
   decodeErrorMessage,
   formatTransactionId,
@@ -44,29 +43,6 @@ import constants from '../../src/lib/constants';
 import { BigNumber as BN } from 'bignumber.js';
 
 describe('Formatters', () => {
-  describe('addPercentageBufferToGasPrice', () => {
-    afterEach(() => {
-      process.env.GAS_PRICE_PERCENTAGE_BUFFER = '0';
-    });
-
-    const TW_COEF = constants.TINYBAR_TO_WEIBAR_COEF;
-    const TEST_CASES = [
-      { testName: 'zero input', buffer: '0', input: 0, output: 0 },
-      { testName: 'buffer 0%', buffer: '0', input: 10 * TW_COEF, output: 10 * TW_COEF },
-      { testName: 'buffer 7%', buffer: '7', input: 140 * TW_COEF, output: 150 * TW_COEF },
-      { testName: 'buffer 10%', buffer: '10', input: 126 * TW_COEF, output: 139 * TW_COEF },
-      { testName: 'buffer 12.25%', buffer: '12.25', input: 56 * TW_COEF, output: 63 * TW_COEF },
-      { testName: 'negative buffer -6%', buffer: '-6', input: 100 * TW_COEF, output: 94 * TW_COEF },
-      { testName: 'negative buffer -12.58%', buffer: '-12.58', input: 136 * TW_COEF, output: 119 * TW_COEF },
-    ];
-    for (let i in TEST_CASES) {
-      it(TEST_CASES[i].testName, () => {
-        process.env.GAS_PRICE_PERCENTAGE_BUFFER = TEST_CASES[i].buffer;
-        expect(addPercentageBufferToGasPrice(TEST_CASES[i].input)).to.equal(TEST_CASES[i].output);
-      });
-    }
-  });
-
   describe('hexToASCII', () => {
     const inputs = ['4C6F72656D20497073756D', '466F6F', '426172'];
